requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
chardet==3.0.4
click==8.0.4
colorama==0.4.4
docutils==0.18.1
Faker==13.3.1
filelock==3.0.12
Flask==2.0.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==2.8
itsdangerous==2.1.0
Jinja2==3.0.3
lockfile==0.12.2
MarkupSafe==2.1.0
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mysqlclient==2.1.0
packaging==21.3
pyparsing==3.0.7
python-daemon==2.2.4
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.22.0
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.25.11
Werkzeug==2.0.3
